S4 Class ISA, initialization method

An S4 class to store information from an ISA-Tab data set, including an investigation file, one or more study files, and one or more assay files for each study file.

When creating a new object of class ISA via: object <- new(Class = "ISA", path), the function initialize(.Object, path) is called to initialize and create the actual object. The initialize-method is seldomly used as a function itself.

# S4 method for ISA
initialize(.Object, path)

Arguments

.Object
character, name of the object of class ISA to be initialized

path
length-one character vector containing the path to the ISA-Tab files of the dataset.

Slots

path

A length-one character vector containing the path to the ISA-Tab dataset.

iFileName

A length-one character vector containing the investigation filename (by definition starting with i_ and ending at .txt).

oSR

A data.frame containing the "ONTOLOGY SOURCE REFERENCE" section of the investigation file.

invest

A data.frame containing the "INVESTIGATION" section of the investigation file.

iPubs

A data.frame containing the "INVESTIGATION PUBLICATIONS" section of the investigation file.

iContacts

A data.frame containing the "INVESTIGATION CONTACTS" section of the investigation file.

study

A list of data.frames containing the "STUDY" sections of the investigation file. Each study has its own section and the study identifier is used to name each element in the list.

sDD

A list of data.frames containing the "STUDY DESIGN DESCRIPTORS" sections of the investigation file. Each study has its own section and the study identifier is used to name each element in the list.

sPubs

A list of data.frames containing the "STUDY PUBLICATIONS" sections of the investigation file. Each study has its own section and the study identifier is used to name each element in the list.

sFacts

A list of data.frames containing the "STUDY FACTORS" sections of the investigation file. Each study has its own section and the study identifier is used to name each element in the list.

sAssays

A list of data.frames containing the "STUDY ASSAYS" sections of the investigation file. Each study has its own section and the study identifier is used to name each element in the list.

sProts

A list of data.frames containing the "STUDY PROTOCOLS" sections of the investigation file. Each study has its own section and he study identifier is used to name each element in the list.

sContacts

A list of data.frames containing the "STUDY CONTACTS" sections of the investigation file. Each study has its own section and the study identifier is used to name each element in the list.

sFiles

A list of data.frames containing the "Study Files", containing the contents of the Study Table files belonging to the studies in the investigation. Each study has one Study Table file.

aFiles

A list of data.frames containing the "Assay Files", containing the contents of the Assay Table files belonging to the studies in the investigation. Each study can have multiple Assay Table files. Each element of the list is named by the Assay File Name.