Get and set oSR

Get and set the ontology Source Reference (oSR) data.frame in an object of ISA.

oSR(x)

oSR(x) <- value

# S4 method for ISA
oSR(x)

# S4 method for ISA
oSR(x) <- value

Arguments

x
An object of class ISA.

value
A data.frame containing the ontology Source Reference information. data.frame in which at least the following columns are present: Term Source Name, Term Source File, Term Source Version, Term Source Description.

Value

A data.frame containing the ontology Source Reference information.

The updated object of class ISA.

Examples

## Read example Atwell data set.
isaObject1 <- readISATab(path = file.path(system.file("extdata/Atwell",
                                          package = "isatabr")))

## Extract ontology Source Reference data.frame.
oSR(isaObject1)

#>   Term Source Name                                 Term Source File
#> 1              OBI       http://data.bioontology.org/ontologies/OBI
#> 2              EFO       http://data.bioontology.org/ontologies/EFO
#> 3               UO                http://purl.obolibrary.org/obo/UO
#> 4        NCBITaxon http://data.bioontology.org/ontologies/NCBITAXON
#> 5               PO        http://data.bioontology.org/ontologies/PO
#> 6              GMI                      http://gwas.gmi.oeaw.ac.at/
#>   Term Source Version
#> 1                  23
#> 2                 118
#> 3                <NA>
#> 4                   6
#> 5                  10
#> 6                <NA>
#>                                                          Term Source Description
#> 1                                         Ontology for Biomedical Investigations
#> 2                                                   Experimental Factor Ontology
#> 3                                                                  Unit Ontology
#> 4 National Center for Biotechnology Information (NCBI) Organismal Classification
#> 5                                                                 Plant Ontology
#> 6                                     Cataloque of Arabidopsis accessions at GMI