Process assay tab data

Process data from assay tab files

Process data from assay tab files with technology type mass spectrometry (ms). Processing those files requires the xcms package to be installed.

Process data from assay tab files with technology type DNA microarray (ms). Processing those files requires the Biobase and affy packages to be installed.

processAssay(isaObject, aTabObject, type = c("raw", "derived"))

# S4 method for ISA,assayTab,character
processAssay(isaObject, aTabObject, type = c("raw", "derived"))

# S4 method for ISA,msAssayTab,character
processAssay(isaObject, aTabObject, type = c("raw", "derived"))

# S4 method for ISA,microarrayAssayTab,character
processAssay(isaObject, aTabObject, type = c("raw", "derived"))

Arguments

isaObject
An object of the ISA.

aTabObject
An object of the microarrayAssayTab.

type
A character string indicating which data files should be processed, either "raw" for raw data files, or "derived" for derived data files. The file names are taken from the corresponding column in the aTabObject.

Examples

### Atwell data.

## Read example Atwell data set.
isaObject1 <- readISATab(path = file.path(system.file("extdata/Atwell",
                                          package = "isatabr")))

## Get assay tabs for isaObject1.
aTabObjects <- getAssayTabs(isaObject1)

## Process assay data.
isaDat <- processAssay(isaObject = isaObject1,
                      aTabObject = aTabObjects$s_study1.txt$a_study1.txt,
                       type = "derived")

## Display first rows and columns.
head(isaDat[, 1:10])

#>   Assay Name      LD  LDV      SD     SDV FT10 FT16 FT22 Seed Dormancy Emco5
#> 1   assay152 6.84105 32.6 93.0417 4.97494   74   87   89            NA    NA
#> 2   assay279      NA   NA      NA      NA   NA   NA   NA            NA    NA
#> 3   assay211      NA   NA      NA      NA   NA   NA   NA            NA    NA
#> 4   assay256      NA   NA      NA      NA   NA   NA   NA            NA    NA
#> 5   assay907      NA   NA      NA      NA   NA   NA   NA            NA    NA
#> 6   assay948      NA   NA      NA      NA   NA   NA   NA            NA    NA

# \donttest{
### faahKO data. - This requires the xcms package to be installed.
## Read ISA-Tab files for faahKO.
if (requireNamespace("xcms")) {
  isaObject3 <- readISATab(path = file.path(system.file("extdata/faahKO",
                                            package = "isatabr")))

  ## Get assay tabs for isaObject3.
  aTabObjects3 <- getAssayTabs(isaObject3)

  ## Process assay data.
  isaDat3 <-
    processAssay(isaObject = isaObject3,
                 aTabObject = aTabObjects3$s_Proteomic_profiling_of_yeast.txt$a_metabolite.txt,
                 type = "raw")

  ## Display output.
  isaDat3
  }

#> Loading required namespace: xcms

# }