Read an ISA-Tab dataset into an R object.

Reads an ISA-Tab dataset, given as a zip file or as a set of files in a specific folder, and builds an object of the ISA class.

readISATab(path = getwd(), zipfile = NULL, verbose = FALSE)

Arguments

path
A character vector with the name of the directory in which the ISA-Tab files are located (if the parameter zipfile is not provided or if it is equal to NULL), or the name of the directory where the zip archive containing the ISA-Tab files is located (if the parameter zipfile is not NULL). The default value is the current working directory.

zipfile
A character vector with the name of the zip archive containing ISA-Tab files themselves (without a directory name in the zip archive). The default value is NULL (specifying that the ISA-Tab files have not been archived in one zipped file).

verbose
Should the messages for the different reading steps be shown?

Value

An object of the ISA class.

Examples

## Read example Atwell data set.
isaObject1 <- readISATab(path = file.path(system.file("extdata/Atwell",
                                          package = "isatabr")))

## Read example zipped Atwell data set.
isazip <- "Atwell.zip"
isaObject2 <- readISATab(path = file.path(system.file("extdata",
                                                      package = "isatabr")),
                         zipfile = isazip,
                         verbose = TRUE)

#> Unzipping file in temporary directory: /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpjgEMo6

#> Unzipped files: a_study1.txt, i_Investigation.txt, s_study1.txt, tdf.txt

#> Converting ISA-Tab dataset at /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpjgEMo6 into an R object...

#> ... done.