Check the validity of an object of class ISA.

The validISAObject function checks whether an object of class ISA is a valid object. An object of the ISA is considered valid when:

There is only one investigation file, which name starts with i_ and ends with .txt, present in the folder containing the ISA-Tab files.

All sections in the investigation file contain at least the columns specified for that specific section at https://isa-specs.readthedocs.io/en/latest/isatab.html

The study and assay files contain at least the columns specified at https://isa-specs.readthedocs.io/en/latest/isatab.html

validISAObject(object)

Arguments

object
An object of class ISA.

Value

TRUE or an error message.

See also

Examples

## Read example Atwell data set.
isaObject1 <- readISATab(path = file.path(system.file("extdata/Atwell",
                                          package = "isatabr")))

## Check validity of the object.
validISAObject(isaObject1)